+++ winsvc.tests/ServiceControlManagerTests.cs
@@ -53,7 +53,7 @@ namespace winsvc.tests
         {
             // Again just checking for a lack of exceptions at this stage
             using (var scm = ServiceControlManager.OpenServiceControlManager(null, SCM_ACCESS.SC_MANAGER_CONNECT))
-            using (scm.OpenService("Spooler", (UInt32) SERVICE_ACCESS.SERVICE_QUERY_STATUS))
+            using (scm.OpenService("Spooler", SERVICE_ACCESS.SERVICE_QUERY_STATUS))
             {
                 // Service is cleaned up in TearDown
             }
@@ -70,7 +70,7 @@ namespace winsvc.tests
             using (var scm = ServiceControlManager.OpenServiceControlManager(null, SCM_ACCESS.SC_MANAGER_CONNECT))
             {
                 // ReSharper disable once AccessToDisposedClosure
-                Assert.That(() => scm.OpenService("Non existant service name", (UInt32) SERVICE_ACCESS.SERVICE_QUERY_STATUS), 
+                Assert.That(() => scm.OpenService("Non existant service name", SERVICE_ACCESS.SERVICE_QUERY_STATUS), 
                     Throws.TypeOf<Win32Exception>()
                           .With.Property("NativeErrorCode").EqualTo(ERROR_SERVICE_DOES_NOT_EXIST));
             }
